RDFKIT-67 All kinds of upgrades but fails at startup with Argot library
Need to replace Argot library with shopt library
