+++ README.md
@@ -1,26 +1,26 @@
 ![EDM Council Logo](etc/image/edmc-logo.jpg)
 
-# rdf-serializer
+# rdf-toolkit
 
-The `rdf-serializer` is a command-line 'swiss army knife' tool for reading and writing RDF and OWL files in whatever format.
+The `rdf-toolkit` is a command-line 'swiss army knife' tool for reading and writing RDF and OWL files in whatever format.
 
-The primary reason for creating this tool was to have a reference implementation of the serializer/formatter that 
+The primary reason for creating this tool was to have a reference implementation of the toolkit/formatter that 
 creates the FIBO ontologies as they are stored in the [Github FIBO repository](https://github.com/edmcouncil/fibo) 
 (which is at this point in time still a private repository). However, this tool is not in any way specific to FIBO, 
 it can be used with any set of ontologies or for that matter even "normal" RDF files.
 
-It currently uses OWLAPI and Sesame to do the hard work, see [this page](docs/dependencies.md) for more info about those products.
+It currently uses OWLAPI and RDF4J to do the hard work, see [this page](docs/dependencies.md) for more info about those products.
 
 This will be used in a commit-hook to make sure that all RDF files in the repo are stored in the same way.
 
-See for more information about developing rdf-serializer [this](docs/develop.md) page or [this page](docs/dependencies.md) for information about dependencies.
+See for more information about developing rdf-toolkit [this](docs/develop.md) page or [this page](docs/dependencies.md) for information about dependencies.
 
 # Recommended Output Format
 
 The recommended Output Format at this time is RDF/XML because that is the format that the OMG requires for submissions. 
-The EDM Council develops the FIBO Ontologies and submits them as RDF/XML, serialized by the `rdf-serializer` to the OMG. 
+The EDM Council develops the FIBO Ontologies and submits them as RDF/XML, serialized by the `rdf-toolkit` to the OMG. 
 So that is why we also use RDF/XML in Github itself. There are some issues with that and we're working on resolving that, 
-by either "fixing" the RDF/XML output generated by OWLAPI or Sesame, or by eventually migrating to some other format. 
+by either "fixing" the RDF/XML output generated by OWLAPI or RDF4J, or by eventually migrating to some other format. 
 For use in Git we need a format that:
 
 ## Requirements for Git-based Ontology Serialization
@@ -33,27 +33,27 @@ For use in Git we need a format that:
 
 # issues
 
-The FIBO JIRA server has a separate project for the rdf-serializer: https://jira.edmcouncil.org/browse/RDFSER
+The FIBO JIRA server has a separate project for the rdf-toolkit: https://jira.edmcouncil.org/browse/RDFSER
 
 Please add your issues, bugs, feature requests, requirements or questions as issues on the JIRA site.
 
 # download
 
-Download the RDF Serializer [here](https://jenkins.edmcouncil.org/job/rdf-serializer-build/lastSuccessfulBuild/artifact/target/scala-2.11/rdf-serializer.jar)
+Download the RDF Toolkit [here](https://jenkins.edmcouncil.org/job/rdf-toolkit-build/lastSuccessfulBuild/artifact/target/scala-2.11/rdf-toolkit.jar)
 
 # usage
 
-Copy the [rdf-serializer.jar](https://jenkins.edmcouncil.org/job/rdf-serializer-build/lastSuccessfulBuild/artifact/target/scala-2.11/rdf-serializer.jar) file to your local disk.
+Copy the [rdf-toolkit.jar](https://jenkins.edmcouncil.org/job/rdf-toolkit-build/lastSuccessfulBuild/artifact/target/scala-2.11/rdf-toolkit.jar) file to your local disk.
 
 ## Linux or Mac OS X
 
-On Linux or Mac OS X you can execute the rdf-serializer
+On Linux or Mac OS X you can execute the rdf-toolkit
 as follows:
 
 1. Open a Terminal
-2. Type the name of the rdf-serializer.jar file on the command prompt and supply the --help option:
+2. Type the name of the rdf-toolkit.jar file on the command prompt and supply the --help option:
 ```
->rdf-serializer.jar --help
+>rdf-toolkit.jar --help
 ```
 
 ## Windows
@@ -61,9 +61,9 @@ as follows:
 1. Open a Command Shell by going to the Start menu and type cmd.exe
 2. Ensure that Java is installed by typing "java -version" on the command line, which should result in
    either version 1.7 or 1.8.
-3. Then launch the rdf-serializer's help function as follows:
+3. Then launch the rdf-toolkit's help function as follows:
 ```
-C:/>java -jar rdf-serializer.jar --help
+C:/>java -jar rdf-toolkit.jar --help
 ```
 
 # --help
@@ -71,16 +71,16 @@ C:/>java -jar rdf-serializer.jar --help
 The current "--help" option gives the following information:
 
 ```
-rdf-serializer version 1.0.3-SNAPSHOT-6d7165e (2015-04-03T20:59:15.113-0400)
+rdf-toolkit version 1.0.3-SNAPSHOT-6d7165e (2015-04-03T20:59:15.113-0400)
 
-Usage: rdf-serializer [OPTIONS] output-file input-file(s) ...
+Usage: rdf-toolkit [OPTIONS] output-file input-file(s) ...
 
 OPTIONS
 
 -h
 --help                                        show this help
 
---version                                     show just the version of rdf-serializer: 1.0.3-SNAPSHOT-6d7165e
+--version                                     show just the version of rdf-toolkit: 1.0.3-SNAPSHOT-6d7165e
 
 -q
 -v
