+++ assembly.sbt
@@ -1,8 +1,8 @@
 //
-// The sbt-assembly specific commands which can be used to construct an "uber-jar" (rdf-serializer.jar) that contains
-// everything to run the RDF serializer from the command line like this:
+// The sbt-assembly specific commands which can be used to construct an "uber-jar" (rdf-toolkit.jar) that contains
+// everything to run the RDF toolkit from the command line like this:
 //
-// java -jar rdf-serializer.jar
+// java -jar rdf-toolkit.jar
 //
 // https://github.com/sbt/sbt-assembly
 //
@@ -18,7 +18,7 @@
 //
 // The name of the uber-jar
 //
-assemblyJarName in assembly := "rdf-serializer.jar"
+assemblyJarName in assembly := "rdf-toolkit.jar"
 
 //
 // To skip the test during assembly:
@@ -28,7 +28,7 @@ test in assembly := {}
 //
 // To set an explicit main class:
 //
-mainClass in assembly := Some("org.edmcouncil.rdf_serializer.Main")
+mainClass in assembly := Some("org.edmcouncil.main.Main")
 
 //
 // Prepend the following shell script to the jar:
@@ -43,22 +43,22 @@ assemblyOption in assembly ~= { _.copy(prependShellScript = Some(shellScript)) }
 //
 // Set the merge strategy for duplicates:
 //
-assemblyMergeStrategy in assembly <<= (assemblyMergeStrategy in assembly) {
-  (old) => {
-    case PathList("javax", "servlet", xs @ _*)         	=> MergeStrategy.first
-    /*
-    case PathList("META-INF", "maven", "com.google.guava", "guava", xs @ _*) => MergeStrategy.first
-    case PathList("META-INF", "maven", "com.fasterxml.jackson.core", "jackson-core", xs @ _*) => MergeStrategy.last
-    case PathList("META-INF", "maven", "commons-codec", "commons-codec", xs @ _*) => MergeStrategy.last
-    case PathList("META-INF", "maven", "org.openrdf.sesame", xs @ _*) => MergeStrategy.last
-    */
-    case PathList("META-INF", "maven", xs @ _*) => MergeStrategy.discard
-    case PathList("META-INF", "sun-jaxb.episode", xs @ _*) => MergeStrategy.first
-    case PathList("org", "apache", "commons", "logging", xs @ _*) => MergeStrategy.first
-    case PathList("com", xs @ _*) => MergeStrategy.last
-    case PathList("info", xs @ _*) => MergeStrategy.last
-    case PathList("org", xs @ _*) => MergeStrategy.last
-    case "booter.properties" => MergeStrategy.first
-    case x => old(x)
-  }
+assemblyMergeStrategy in assembly := {
+  case PathList("javax", "servlet", xs @ _*)         	=> MergeStrategy.first
+  /*
+  case PathList("META-INF", "maven", "com.google.guava", "guava", xs @ _*) => MergeStrategy.first
+  case PathList("META-INF", "maven", "com.fasterxml.jackson.core", "jackson-core", xs @ _*) => MergeStrategy.last
+  case PathList("META-INF", "maven", "commons-codec", "commons-codec", xs @ _*) => MergeStrategy.last
+  case PathList("META-INF", "maven", "org.openrdf.sesame", xs @ _*) => MergeStrategy.last
+  */
+  case PathList("META-INF", "maven", xs @ _*) => MergeStrategy.discard
+  case PathList("META-INF", "sun-jaxb.episode", xs @ _*) => MergeStrategy.first
+  case PathList("org", "apache", "commons", "logging", xs @ _*) => MergeStrategy.first
+  case PathList("com", xs @ _*) => MergeStrategy.last
+  case PathList("info", xs @ _*) => MergeStrategy.last
+  case PathList("org", xs @ _*) => MergeStrategy.last
+  case "booter.properties" => MergeStrategy.first
+  case x =>
+    val oldStrategy = (assemblyMergeStrategy in assembly).value
+    oldStrategy(x)
 }
