+++ docs/dependencies.md
@@ -1,18 +1,18 @@
 # Dependencies
 
-The rdf-serializer works with the following major libraries:
+The rdf-toolkit works with the following major libraries:
 
 # [OWLAPI](https://github.com/owlcs/owlapi)
 
 - Sources - https://github.com/owlcs/owlapi
 - Docs - http://owlapi.sourceforge.net/documentation.html
 
-# [Sesame](https://bitbucket.org/openrdf/sesame)
+# [RDF4J](http://rdf4j.org/)
 
-- Sources - https://bitbucket.org/openrdf/sesame or https://github.com/ansell/openrdf-sesame
-- Docs - http://rdf4j.org/documentation.docbook?view
+- Sources - https://github.com/eclipse/rdf4j
+- Docs - http://docs.rdf4j.org/
 
-The Sesame API is embedded in the code (currently as Java code) and can be invoked in two ways:
+The RDF4J API is embedded in the code (currently as Java code) and can be invoked in two ways:
 
 - Through its own Main Class from the command line
-- Through the primary command line interface via the `--api sesame` switch.
+- Through the primary command line interface via the `--api rdf4j` switch.
