+++ project/plugins.sbt
@@ -1,6 +1,6 @@
 
 
-addSbtPlugin("com.eed3si9n" % "sbt-assembly" % "0.13.0")
+addSbtPlugin("com.eed3si9n" % "sbt-assembly" % "0.14.3")
 
 //
 // https://github.com/orrsella/sbt-sublime
